04/06 fin de journée

+++ src/main/java/repository/FormateurRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 import model.Apprenant;
 import model.Formateur;
 
-public interface FormateurRepository extends CrudRepository<Formateur, Long>{
+public interface FormateurRepository extends CrudRepository<Formateur, Integer>{
